Language.LSP.Logging

Documentation

logToShowMessage :: MonadLsp c m => LogAction m (WithSeverity Text)

Logs messages to the client via window/showMessage.

logToLogMessage :: MonadLsp c m => LogAction m (WithSeverity Text)

Logs messages to the client via window/logMessage.

defaultClientLogger :: MonadLsp c m => LogAction m (WithSeverity Text)

A sensible log action for logging messages to the client:

Shows Error logs to the user via window/showMessage

Logs Info and above logs in the client via window/logMessage

If you want finer control (e.g. the ability to log Debug logs based on a flag, or similar), then do not use this and write your own based on logToShowMessage and logToLogMessage.