Language.LSP.VFS

Description

Handles the Language.LSP.Types.TextDocumentDidChange / Language.LSP.Types.TextDocumentDidOpen / Language.LSP.Types.TextDocumentDidClose messages to keep an in-memory filesystem of the current client workspace. The server can access and edit files in the client workspace by operating on the VFS in LspFuncs.

Documentation

data VFS

Constructors

VFS

Fields

_vfsMap :: !(Map NormalizedUri VirtualFile)

vfsMap :: HasVfsMap s a => Lens' s a

data VirtualFile

Constructors

VirtualFile

Fields

_lsp_version :: !Int32

The LSP version of the document

_file_version :: !Int

This number is only incremented whilst the file remains in the map.

_file_text :: !Rope

The full contents of the document

lsp_version :: HasLsp_version s a => Lens' s a

file_version :: HasFile_version s a => Lens' s a

file_text :: HasFile_text s a => Lens' s a

virtualFileText :: VirtualFile -> Text

virtualFileVersion :: VirtualFile -> Int32

data VfsLog

Constructors

SplitInsideCodePoint Position Rope

URINotFound NormalizedUri

Opening NormalizedUri

Closing NormalizedUri

PersistingFile NormalizedUri FilePath

CantRecursiveDelete NormalizedUri

DeleteNonExistent NormalizedUri

Managing the VFS

emptyVFS :: VFS

openVFS :: MonadState VFS m => LogAction m (WithSeverity VfsLog) -> TMessage 'Method_TextDocumentDidOpen -> m ()

Applies the changes from a DidOpenTextDocument to the VFS

changeFromClientVFS :: MonadState VFS m => LogAction m (WithSeverity VfsLog) -> TMessage 'Method_TextDocumentDidChange -> m ()

Applies a DidChangeTextDocumentNotification to the VFS

changeFromServerVFS :: forall m. MonadState VFS m => LogAction m (WithSeverity VfsLog) -> TMessage 'Method_WorkspaceApplyEdit -> m ()

Applies the changes from a ApplyWorkspaceEditRequest to the VFS

persistFileVFS :: MonadIO m => LogAction m (WithSeverity VfsLog) -> FilePath -> VFS -> NormalizedUri -> Maybe (FilePath, m ())

Write a virtual file to a file in the given directory if it exists in the VFS.

closeVFS :: MonadState VFS m => LogAction m (WithSeverity VfsLog) -> TMessage 'Method_TextDocumentDidClose -> m ()

Positions and transformations

data CodePointPosition

A position, like a Position, but where the offsets in the line are measured in Unicode code points instead of UTF-16 code units.

Constructors

CodePointPosition

Fields

_line :: UInt

Line position in a document (zero-based).

_character :: UInt

Character offset on a line in a document in *code points* (zero-based).

line :: HasLine s a => Lens' s a

character :: HasCharacter s a => Lens' s a

codePointPositionToPosition :: VirtualFile -> CodePointPosition -> Maybe Position

Given a virtual file, translate a CodePointPosition in that file into a Position in that file.

Will return Nothing if the requested position is out of bounds of the document.

Logarithmic in the number of lines in the document, and linear in the length of the line containing the position.

positionToCodePointPosition :: VirtualFile -> Position -> Maybe CodePointPosition

Given a virtual file, translate a Position in that file into a CodePointPosition in that file.

Will return Nothing if the requested position lies inside a code point, or if it is out of bounds of the document.

Logarithmic in the number of lines in the document, and linear in the length of the line containing the position.

data CodePointRange

A range, like a Range, but where the offsets in the line are measured in Unicode code points instead of UTF-16 code units.

Constructors

CodePointRange

Fields

_start :: CodePointPosition

The range's start position.

_end :: CodePointPosition

The range's end position.

start :: HasStart s a => Lens' s a

end :: HasEnd s a => Lens' s a

codePointRangeToRange :: VirtualFile -> CodePointRange -> Maybe Range

Given a virtual file, translate a CodePointRange in that file into a Range in that file.

Will return Nothing if any of the positions are out of bounds of the document.

Logarithmic in the number of lines in the document, and linear in the length of the lines containing the positions.

rangeToCodePointRange :: VirtualFile -> Range -> Maybe CodePointRange

Given a virtual file, translate a Range in that file into a CodePointRange in that file.

Will return Nothing if any of the positions are out of bounds of the document.

Logarithmic in the number of lines in the document, and linear in the length of the lines containing the positions.

manipulating the file contents

rangeLinesFromVfs :: VirtualFile -> Range -> Text

data PosPrefixInfo

Describes the line at the current cursor position

Constructors

PosPrefixInfo

Fields

fullLine :: !Text

The full contents of the line the cursor is at

prefixModule :: !Text

If any, the module name that was typed right before the cursor position. For example, if the user has typed "Data.Maybe.from", then this property will be Data.Maybe

prefixText :: !Text

The word right before the cursor position, after removing the module part. For example if the user has typed "Data.Maybe.from", then this property will be "from"

cursorPos :: !Position

The cursor position

getCompletionPrefix :: Monad m => Position -> VirtualFile -> m (Maybe PosPrefixInfo)

for tests

applyChanges :: Monad m => LogAction m (WithSeverity VfsLog) -> Rope -> [TextDocumentContentChangeEvent] -> m Rope

Apply the list of changes. Changes should be applied in the order that they are received from the client.

applyChange :: Monad m => LogAction m (WithSeverity VfsLog) -> Rope -> TextDocumentContentChangeEvent -> m Rope

changeChars :: Monad m => LogAction m (WithSeverity VfsLog) -> Rope -> Position -> Position -> Text -> m Rope

Given a Rope, start and end positions, and some new text, replace the given range with the new text. If the given positions lie within a code point then this does nothing (returns the original Rope) and logs.